+++ tests/testthat/test-eunomia-refit-ps.R
@@ -4,7 +4,9 @@ library(testthat)
 test_that("Multiple analyses", {
   tcos1 <- createTargetComparatorOutcomes(targetId = 1,
                                           comparatorId = 2,
-                                          outcomeIds = c(3, 4),
+                                          outcomes = list(createOutcome(outcomeId = 3,
+                                                                        priorOutcomeLookback = 30),
+                                                          createOutcome(outcomeId = 4)),
                                           excludedCovariateConceptIds = c(1118084, 1124300))
   targetComparatorOutcomesList <- list(tcos1)